feat(php): add php_openads native PHP extension for OpenADS
Adds bindings/php_ext/ — a ZTS x64 C extension (php_openads.dll) that
exposes the OpenADS ACE API to PHP 8.0 with the same class surface as
php_advantage, enabling drop-in use by the DataArchitect web app.

Classes: AdsConnection, AdsStatement, AdsTable, AdsDictionary,
AdsTransaction, AdsPreparedStatement, AdsException.

Build: Makefile.win + build.bat (VS2022 x64, links against openace64.lib).
The engine DLL is shipped as openace64.dll (a rename of ace64.dll) to
avoid collision with the SAP ace64.dll that php_ads.dll depends on.

Also fixes connection.cpp: SQL queries with an explicit .adt or .ntx
extension (e.g. SELECT * FROM tbl.adt) now open the file with the
correct driver instead of falling through to the CDX/DBF default.

+++ src/engine/data_dict.cpp
@@ -13,16 +13,17 @@ namespace openads::engine {
 namespace {
 
 static inline uint32_t le32(const std::string& b, std::size_t off) {
-    return static_cast<uint32_t>(static_cast<uint8_t>(b[off]))
-         | (static_cast<uint32_t>(static_cast<uint8_t>(b[off+1])) << 8)
-         | (static_cast<uint32_t>(static_cast<uint8_t>(b[off+2])) << 16)
-         | (static_cast<uint32_t>(static_cast<uint8_t>(b[off+3])) << 24);
+    return static_cast<uint32_t>(
+        static_cast<uint32_t>(static_cast<uint8_t>(b[off]))
+      | (static_cast<uint32_t>(static_cast<uint8_t>(b[off+1])) << 8)
+      | (static_cast<uint32_t>(static_cast<uint8_t>(b[off+2])) << 16)
+      | (static_cast<uint32_t>(static_cast<uint8_t>(b[off+3])) << 24));
 }
 
 static inline uint16_t le16(const std::string& b, std::size_t off) {
     return static_cast<uint16_t>(
-        static_cast<unsigned>(static_cast<uint8_t>(b[off]))
-      | (static_cast<unsigned>(static_cast<uint8_t>(b[off+1])) << 8));
+        static_cast<uint16_t>(static_cast<uint8_t>(b[off]))
+      | static_cast<uint16_t>(static_cast<uint16_t>(static_cast<uint8_t>(b[off+1])) << 8));
 }
 
 static void put_le32(std::string& b, std::size_t off, uint32_t v) {
